phone_directory/nlg/dev-requirements.txt
pytest==5.4.1
pytest-flask==0.15.1
